+++ CHANGELOG.md
@@ -4,6 +4,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.9.0] - 2023-03-29
+### Added
+- Tree Modifier: Ability to copy/shift nodes with full path in `from_paths` for faster search performance, added `with_full_path` parameter.
+### Changed
+- Tree Modifier: Enforced paths in `to_paths` to be full path for faster search performance.
+- Tree Modifier: Faster creation of intermediate parent nodes in `to_paths`.
+- Tree Modifier: Better handling of `sep` in paths by performing string replacement at the start.
+- Tree Modifier: Check and throw error for invalid parameters, case when node is meant to be deleted but `copy=True`.
+### Fixed
+- Tree Modifier: Fix issue trailing `sep` differing in `from_paths` and `to_paths` which should not throw error.
+
 ## [0.8.4] - 2023-03-24
 ### Added
 - Tree Search: Implement `find_child` and `find_children` to find single child or multiple children based on user-defined condition.
@@ -236,6 +247,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Utility Iterator: Tree traversal methods.
 - Workflow To Do App: Tree use case with to-do list implementation.
 
+[0.9.0]: https://github.com/kayjan/bigtree/compare/v0.8.4...v0.9.0
 [0.8.4]: https://github.com/kayjan/bigtree/compare/v0.8.3...v0.8.4
 [0.8.3]: https://github.com/kayjan/bigtree/compare/v0.8.2...v0.8.3
 [0.8.2]: https://github.com/kayjan/bigtree/compare/v0.8.1...v0.8.2